fix(autoStart): use shared SM encoded config builder
autoStart.triggerConfiguredWorkflowForTicket previously built a minimal encoded_config containing only inputJql and customParams. This caused ai-teammate.yml to fall back to a hard-coded list of common prompts and lose the target agent JSON's full cliPrompts.

Extract buildEncodedConfig from smAgent.js into agents/js/common/buildEncodedConfig.js and use it from both SM Agent and autoStart. The encoded config now always carries the resolved agent params: cliPrompts, customParams, feedbackLoop, managedSubmodules, etc.
